+++ package.json
@@ -8,31 +8,32 @@
     "test": "test"
   },
   "bin": {
-    "musicxml-mma": "./server.js",
-    "ireal-midi": "./scripts/ireal-midi.js",
-    "midi-timemap": "./scripts/midi-timemap.js"
+    "musicxml-mma": "src/js/server.js",
+    "ireal-midi": "src/js/ireal-midi.js",
+    "midi-timemap": "src/js/midi-timemap.js"
   },
   "scripts": {
     "build:mmarc": "cp mmarc.example mmarc",
-    "build:grooves": "${MMA_HOME:-./mma}/mma.py -G && npm run --silent print:grooves > grooves.txt",
-    "build:sef": "xsls=('musicxml-unroll' 'musicxml-mma-unrolled' 'musicxml-mma' 'musicxml-timemap'); for xsl in ${xsls[@]}; do xslt3 -relocate:on -xsl:$xsl.xsl -export:$xsl.sef.json -nogo:1 -t -ns:##html5; done && rm -f cache/*.mid",
+    "build:grooves": "${MMA_HOME:-mma}/mma.py -G && npm run --silent print:grooves > build/grooves.txt",
+    "build:sef": "xsls=('musicxml-unroll' 'musicxml-mma-unrolled' 'musicxml-mma' 'musicxml-timemap'); for xsl in ${xsls[@]}; do xslt3 -relocate:on -xsl:src/xsl/$xsl.xsl -export:./build/$xsl.sef.json -nogo:1 -t -ns:##html5; done && rm -f cache/*.mid",
     "build": "[ ! -f mmarc ] && npm run build:mmarc; npm run build:grooves; npm run build:sef",
-    "convert:unroll": "run() { xslt3 -xsl:musicxml-unroll.xsl -s:\"$1\" -o:\"${1/.musicxml/.unrolled.musicxml}\" ${@:2}; }; run",
-    "convert:timemap": "run() { xslt3 -xsl:musicxml-timemap.xsl -s:\"$1\" -o:\"${1/.musicxml/.timemap.json}\" ${@:2}; }; run",
-    "convert:mma": "run() { xslt3 -xsl:musicxml-mma.xsl -s:\"$1\" -o:\"${1/.musicxml/.mma}\" ${@:2}; }; run",
-    "convert:midi": "run() { ${MMA_HOME:-./mma}/mma.py -II \"$1\" -f \"${1/.mma/.mid}\"; }; run",
+    "convert:unroll": "run() { xslt3 -xsl:src/xsl/musicxml-unroll.xsl -s:\"$1\" ${@:2}; }; run",
+    "convert:timemap": "run() { xslt3 -xsl:src/xsl/musicxml-timemap.xsl -s:\"$1\" ${@:2}; }; run",
+    "convert:mma": "run() { xslt3 -xsl:src/xsl/musicxml-mma.xsl -s:\"$1\" ${@:2}; }; run",
+    "convert:midi": "run() { ${MMA_HOME:-mma}/mma.py -II \"$1\" -f \"${1/.mma/.mid}\"; }; run",
     "convert": "run() { npm run convert:mma \"$1\" ${@:2} && npm run convert:midi \"${1/.musicxml/.mma}\"; }; run",
-    "print:chord": "run() { echo \"PrintChord $1\" | ${MMA_HOME:-./mma}/mma.py -n -; }; run",
-    "print:grooves": "find ${MMA_HOME:-./mma}/lib grooves -name '*.mma' | while read f; do MMA_ENCODING=utf-8 ${MMA_HOME:-./mma}/mma.py -Dbo \"$f\" | tail -n +2; done",
-    "print:preview": "run() { ${MMA_HOME:-./mma}/mma.py -V \"$@\"; }; run",
-    "validate:musicxml": "run() { xmllint --noout --schema musicxml.xsd \"$1\"; }; run",
-    "validate:mma": "run() { ${MMA_HOME:-./mma}/mma.py -II -n \"$1\"; }; run",
-    "develop": "nodemon -e js,json server.js",
-    "test:xsl": "./test/libs/bats/bin/bats test/*.bats",
+    "print:chord": "run() { echo \"PrintChord $1\" | ${MMA_HOME:-mma}/mma.py -n -; }; run",
+    "print:grooves": "find ${MMA_HOME:-mma}/lib grooves -name '*.mma' | while read f; do MMA_ENCODING=utf-8 ${MMA_HOME:-mma}/mma.py -Dbo \"$f\" | tail -n +2; done",
+    "print:preview": "run() { ${MMA_HOME:-mma}/mma.py -V \"$@\"; }; run",
+    "validate:musicxml": "run() { xmllint --noout --schema src/xsd/musicxml.xsd \"$1\"; }; run",
+    "validate:mma": "run() { ${MMA_HOME:-mma}/mma.py -II -n \"$1\"; }; run",
+    "develop": "nodemon -e js,json src/js/server.js",
+    "start": "node src/js/server.js",
+    "test:xsl": "test/libs/bats/bin/bats test/*.bats",
     "test:server": "cross-env PORT=1331 NODE_OPTIONS=--experimental-vm-modules NODE_ENV=test jest",
     "test": "npm run test:xsl && npm run test:server",
     "server": "PORT=8085 npm run start",
-    "postinstall": "./scripts/postinstall.sh"
+    "postinstall": "./postinstall.sh"
   },
   "repository": {
     "type": "git",
