+++ plugins
@@ -1,0 +1,1 @@
+src/plugins/
