+++ src/grooves/arabic/ayyub.mma
