+++ test/04-musicxml-timemap.bats
@@ -5,7 +5,7 @@ load 'libs/bats-assert/load'
 set -euo pipefail
 
 @test "musicxml-timemap produces a valid and correct JSON file for asa-branca" {
-  timemap=$(xslt3 -xsl:musicxml-timemap.xsl -s:test/data/asa-branca.musicxml)
+  timemap=$(xslt3 -xsl:src/xsl/musicxml-timemap.xsl -s:test/data/asa-branca.musicxml)
   echo "${timemap}" | jq type 1>/dev/null
   select=$(echo "${timemap}" | jq '.[] | select(.measure == 1)')
   run echo ${select}
